+++ src/components/Navbar.js
@@ -3,13 +3,9 @@ import { useSelector } from 'react-redux/es/hooks/useSelector';
 // import { NavLink } from 'react-router-dom';
 
 import backBtn from '../assets/back.png';
+import gearBtn from '../assets/gear.png';
+import micBtn from '../assets/mic.png';
 import styles from '../styles/Navbar.module.css';
-
-const link = [
-  {
-    path: '/', text: 'back',
-  },
-];
 
 const Navbar = () => {
   const { PageName } = useSelector((store) => store.layout);
@@ -17,8 +13,14 @@ const Navbar = () => {
   return (
     <nav>
       <div className={styles.look}>
-        <img src={backBtn} alt={link.text} className={styles.whiteArrow} />
-        <h1>{PageName}</h1>
+        <div className={styles.aligned}>
+          <img src={backBtn} alt="back button" className={styles.whiteImage} />
+          <h1>{PageName}</h1>
+        </div>
+        <div className={styles.aligned}>
+          <img src={micBtn} alt="mic button" className={`${styles.whiteImage} ${styles.options}`} />
+          <img src={gearBtn} alt="settings button" className={`${styles.whiteImage} ${styles.options}`} />
+        </div>
       </div>
     </nav>
   );
