+++ src/styles/home.module.css
@@ -1,3 +1,3 @@
-.hidden {
-  display: none;
+.look {
+  background-color: rgb(235, 91, 168);
 }
